requirements.txt
amqp==5.3.1
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
decorator==5.2.1
Django==5.2.6
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.2
django-jazzmin==3.0.1
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
executing==2.2.1
idna==3.10
inflection==0.5.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
matplotlib-inline==0.1.7
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pydotplus==2.0.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
gunicorn